add handling errors with rolling window/monotonic check to array_utils
