+++ rt_segment_speeds/README.md
@@ -10,6 +10,7 @@ Work related to using `dask` to produce speeds by segments from GTFS RT vehicle 
     * scripts run consecutively within each stage `A1`, `A2`, ...
 * exploratory notebooks: prefix with `00_`, `01_`, `02_`, etc.
 * [Exploratory epic](https://github.com/cal-itp/data-analyses/issues/592)
+* refer to [Makefile](./scripts/Makefile) and [config](./scripts/config.yml) for pipeline
 
 | warehouse         | gcs               | github             |
 |-------------------|-------------------|--------------------|
@@ -19,8 +20,3 @@ Work related to using `dask` to produce speeds by segments from GTFS RT vehicle 
 | rt_predictions    | rt_predictions    | rt_predictions     | 
 
 
-## Scripts
-1. cut segments (route segments or stop segments)
-1. spatial join vehicle positions to the segments  (`A3`)
-1. pare down vehicle positions to just enter/exit within a segment + placeholder for dropping unusable trips. should exclude trips with too little info or `trip_id is None` at the start, then pare down (`A4`)
-1. do linear referencing within segments to get `distance_elapsed` and `time_elapsed`, and calculate speeds. save as partitioned parquets (`A5`)
