+++ rt_segment_speeds/setup.py
@@ -3,7 +3,7 @@ from setuptools import find_packages, setup
 setup(
     name="segment_speed_utils",
     packages=find_packages(),
-    version="1.0.0",
+    version="1.2.0",
     description="Utility functions for GTFS RT segment speeds",
     author="Cal-ITP",
     license="Apache",
